% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterflow.R
\name{waterflow}
\alias{waterflow}
\title{Solve the Waterflow Problem}
\description{
This \code{R6} Class provides methods to calculate the total number of
squares which will fill with water and to plot the walls and filled valleys.
}
\section{Usage}{

\preformatted{
  p <- waterflow$new(wall = NULL)
  p$total()
  p$plot()
  p$print()
}
}

\section{Arguments}{

\describe{
  \item{p}{A \code{waterflow} object.}
  \item{wall}{A numeric vector of wall heights}
}
}

\section{Details}{

\code{$new()} calculates the amount of water the valleys will be filled with.

\code{$total()} returns the total amount of water the valleys will be filled
with.

\code{$plot()} plots the walls filled with water.

\code{$print()} returns the \code{data.frame} of wall and water heights.
}

\examples{
x <- c(2, 5, 1, 2, 3, 4, 7, 7, 6)
p <- waterflow$new(x)
p$total()
p$plot()
p$print()
}
